speculation-rules/prefetch/prefetch-service-worker-fallback.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.sub.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>
  promise_test(async t => {
    assert_implements(HTMLScriptElement.supports('speculationrules'), "Speculation Rules not supported");

    let agent = await spawnWindow(t, { protocol: 'https' });
    let nextUrl = agent.getExecutorURL({ protocol: 'https', page: 2 });

    await agent.forceSinglePrefetch(nextUrl);

    const r = await service_worker_unregister_and_register(t, 'resources/sw.js', nextUrl);
    t.add_cleanup(function() {
      return service_worker_unregister(t, nextUrl);
    });
    await wait_for_state(t, r.installing, 'activated');

    await agent.navigate(nextUrl);

    assert_prefetched(await agent.getRequestHeaders(), "Prefetch should work for HTTPS urls.");
  }, `HTTPS url prefetch consumption should work after service worker fallback`);
</script>
